chatbot-app-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
tensorflow==2.15.0
yfinance==0.2.28
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
python-multipart==0.0.6
openai==1.3.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
alembic==1.13.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
